+++ source/gesture inference/README.md
@@ -1,0 +1,65 @@
+<div align="center">
+
+# Motion
+[![Report Issue on Jira](https://img.shields.io/badge/Report%20Issues-Jira-0052CC?style=flat&logo=jira-software)](https://temple-cis-projects-in-cs.atlassian.net/jira/software/c/projects/DT/issues)
+[![Deploy Docs](https://github.com/ApplebaumIan/tu-cis-4398-docs-template/actions/workflows/deploy.yml/badge.svg)](https://github.com/ApplebaumIan/tu-cis-4398-docs-template/actions/workflows/deploy.yml)
+[![Documentation Website Link](https://img.shields.io/badge/-Documentation%20Website-brightgreen)](https://applebaumian.github.io/tu-cis-4398-docs-template/)
+
+
+</div>
+
+
+## Keywords
+
+Section #, as well as any words that quickly give your peers insights into the application like programming language, development platform, type of application, etc.
+
+## Project Abstract
+
+This document proposes a novel application of a text message (SMS or Email) read-out and hands-free call interacted between an Android Smartphone and an infotainment platform (headunit) in a car environment. When a phone receives an SMS or Email, the text message is transferred from the phone to the headunit through a Bluetooth connection. On the headunit, user can control which and when the received SMS or E-mail to be read out through the in-vehicle audio system. The user may press one button on the headunit to activate the hands-free feature to call back the SMS sender.
+
+## High Level Requirement
+
+Describe the requirements – i.e., what the product does and how it does it from a user point of view – at a high level.
+
+## Conceptual Design
+
+Describe the initial design concept: Hardware/software architecture, programming language, operating system, etc.
+
+## Background
+
+The background will contain a more detailed description of the product and a comparison to existing similar projects/products. A literature search should be conducted and the results listed. Proper citation of sources is required. If there are similar open-source products, you should state whether existing source will be used and to what extent. If there are similar closed-source/proprietary products, you should state how the proposed product will be similar and different.
+
+## Required Resources
+
+Discuss what you need to develop this project. This includes background information you will need to acquire, hardware resources, and software resources. If these are not part of the standard Computer Science Department lab resources, these must be identified early and discussed with the instructor.
+
+## Collaborators
+
+[//]: # ( readme: collaborators -start )
+<table>
+<tr>
+    <td align="center">
+        <a href="https://github.com/ApplebaumIan">
+            <img src="https://avatars.githubusercontent.com/u/9451941?v=4" width="100;" alt="ApplebaumIan"/>
+            <br />
+            <sub><b>Ian Tyler Applebaum</b></sub>
+        </a>
+    </td>
+    <td align="center">
+        <a href="https://github.com/leekd99">
+            <img src="https://avatars.githubusercontent.com/u/32583417?v=4" width="100;" alt="leekd99"/>
+            <br />
+            <sub><b>Kyle Dragon Lee</b></sub>
+        </a>
+    </td>
+    <td align="center">
+        <a href="https://github.com/thanhnguyen46">
+            <img src="https://avatars.githubusercontent.com/u/60533187?v=4" width="100;" alt="thanhnguyen46"/>
+            <br />
+            <sub><b>Thanh Nguyen</b></sub>
+        </a>
+    </td>
+   </tr>
+</table>
+
+[//]: # ( readme: collaborators -end )
