+++ source/gesture inference/Webcam.py
@@ -1,0 +1,28 @@
+# https://github.com/nrsyed/computer-vision/blob/master/multithread/VideoShow.py
+
+import cv2
+class Webcam:
+    """
+    Class that continuously gets frames and extracts hand data
+    with a dedicated thread and Mediapipe
+    """
+    def __init__(
+        self,
+        webcam_id=0,
+    ):
+        # OpenCV setup
+        self.stream = cv2.VideoCapture(webcam_id)
+        # motion JPG format
+        self.stream.set(cv2.CAP_PROP_FOURCC, cv2.VideoWriter_fourcc("M", "J", "P", "G"))
+        (self.grabbed, self.frame) = self.stream.read()
+        self.frame = cv2.flip(self.frame, 1)
+
+    def stop(self):
+        self.stream.release()
+        cv2.destroyAllWindows()
+
+    def read(self, flip=True):
+        (self.grabbed, self.frame) = self.stream.read()
+        if flip:
+            self.frame = cv2.flip(self.frame, 1)
+        return (self.grabbed, self.frame)
