+++ source/gesture inference/test.py
@@ -1,0 +1,21 @@
+import numpy
+
+model_inpout = [[[
+    -0.00355055,  0.08615325,  0.03645849,  0.02479225,  0.06601369,
+    0.01774658,  0.05041692,  0.0534518,   0.00560722,  0.07558247,
+    0.0355779,  -0.00757914,  0.09286362,  0.01831537, -0.01189687,
+    0.02417924, -0.00065195,  0.00143245,  0.03041233, -0.03012565,
+    -0.00574961,  0.03308015, -0.05212945, -0.01117697,  0.0343622,
+    -0.0663949,  -0.03231894,  0.00075055, -0.00597357,  0.00417856,
+    0.00055868, -0.04379579, -0.00400382, -0.00932142, -0.06584374,
+    -0.01890792, -0.00811052, -0.08931509, -0.03368567, -0.01418387,
+    -0.00079544, -0.00214285, -0.02565904, -0.03279686, -0.00806323,
+    -0.03242598, -0.05364299, -0.01863396, -0.03277639, -0.07600781,
+    -0.03362795, -0.03316592,  0.01428213, -0.00483436, -0.04393554,
+    -0.00857823, -0.00672621, -0.05217818, -0.02510658, -0.01156851,
+    -0.0543259,  -0.04207192, -0.02559664, -0.48394126, -0.45773405
+   ]]]
+
+asdf = numpy.asarray(model_inpout)
+print(asdf)
+print(asdf.shape[0], asdf.shape[2])
